Refactored package name, and updated classes with the removal of prepareFailView.
