+++ src/main/java/delete_task_use_case/DeleteTaskController.java
@@ -5,7 +5,7 @@
  * Created: Nov 19, 2022
  * Last Modified: Nov 19, 2022
  */
-package task_deletion_use_case;
+package delete_task_use_case;
 
 public class DeleteTaskController {
     private DeleteTaskInputBoundary inputBoundary;
