+++ src/main/java/delete_task_use_case/DeleteTaskDataAccess.java
@@ -5,7 +5,7 @@
  * Created: Nov 19, 2022
  * Last Modified: Nov 19, 2022
  */
-package task_deletion_use_case;
+package delete_task_use_case;
 
 import data_access.DataAccessDay;
 import data_access.DataAccessTask;
@@ -14,13 +14,6 @@ import data_access.WeekDataAccess;
 import java.util.HashMap;
 
 public class DeleteTaskDataAccess extends WeekDataAccess implements DeleteTaskDsGateway {
-
-    @Override
-    public boolean taskExists(int dayIndex, String taskTitle) {
-        DataAccessDay day = super.getDays().get(dayIndex);
-        return day.getTasks().containsKey(taskTitle);
-    }
-
     @Override
     public void save(DeleteTaskDataAccessInput dataAccessInput) {
         int dayIndex = dataAccessInput.getDayIndex();
