+++ src/main/java/delete_task_use_case/DeleteTaskDataAccessInput.java
@@ -1,18 +1,17 @@
 /**
- * Delete Task InputData Class.
+ * Delete Task DataAccessInput Class.
  * Author: Ricky Fung
  * Modified By: Ricky Fung
  * Created: Nov 19, 2022
  * Last Modified: Nov 19, 2022
  */
-package task_deletion_use_case;
+package delete_task_use_case;
 
-public class DeleteTaskInputData {
-
+public class DeleteTaskDataAccessInput {
     private int dayIndex;
     private String taskTitle;
 
-    public DeleteTaskInputData(int dayIndex, String taskTitle) {
+    public DeleteTaskDataAccessInput(int dayIndex, String taskTitle) {
         this.taskTitle = taskTitle;
         this.dayIndex = dayIndex;
     }
