+++ src/main/java/delete_task_use_case/DeleteTaskDsGateway.java
@@ -5,10 +5,8 @@
  * Created: Nov 19, 2022
  * Last Modified: Nov 19, 2022
  */
-package task_deletion_use_case;
+package delete_task_use_case;
 
 public interface DeleteTaskDsGateway {
-    boolean taskExists(int dayIndex, String taskTitle);
-
     void save(DeleteTaskDataAccessInput dataAccessInput);
 }
