+++ src/main/java/delete_task_use_case/DeleteTaskPresenter.java
@@ -5,16 +5,11 @@
  * Created: Nov 19, 2022
  * Last Modified: Nov 19, 2022
  */
-package task_deletion_use_case;
+package delete_task_use_case;
 
 public class DeleteTaskPresenter implements DeleteTaskOutputBoundary {
     @Override
     public DeleteTaskOutputData prepareSuccessView(DeleteTaskOutputData taskOutputData) {
-        return taskOutputData;
-    }
-
-    @Override
-    public DeleteTaskOutputData prepareFailView(DeleteTaskOutputData taskOutputData) {
         return taskOutputData;
     }
 }
